+++ src/PairingPlus.cpp
@@ -43,7 +43,7 @@
 PairingPlus::PairingPlus(Position arg_position1, Position arg_position2):
 	position1(arg_position1),
 	position2(arg_position2),
-	size(arg_position2 - arg_position1 - 1)
+	size( (size_t) (arg_position2 - arg_position1 - 1) )
 {
 	this->init();
 }
